+++ src/lib/components/Overlay/Overlay.svelte
@@ -9,6 +9,7 @@
 	export { klass as class };
 	export let style = '';
 
+	export let overlay: HTMLDivElement | undefined = undefined;
 	export let transition: <T>(node: HTMLDivElement, options: T) => TransitionConfig = fade;
 	export let inOpts: T = { duration: 250 } as unknown as T;
 	export let outOpts: T = { duration: 250 } as unknown as T;
@@ -18,12 +19,14 @@
 
 {#if active}
 	<div
-		in:transition={inOpts}
-		out:transition={outOpts}
+		bind:this={overlay}
 		class="s-overlay {klass}"
 		class:absolute
+		in:transition={inOpts}
+		out:transition={outOpts}
 		on:click
 		{style}
+		{...$$restProps}
 	>
 		<div class="s-overlay__scrim" />
 		<div class="s-overlay__content">