+++ src/lib/components/Overlay/index.ts
@@ -1,1 +1,13 @@
+import { TransitionConfig } from 'svelte/transition';
+
 export { default } from './Overlay.svelte';
+
+export interface OverlayProps<InTransitionProps, OutTransitionProps> {
+	class?: string;
+	style?: string;
+	inOpts?: InTransitionProps;
+	outOpts?: OutTransitionProps;
+	active?: boolean;
+	absolute?: boolean;
+	transition?<T>(node: HTMLDivElement, options: T): TransitionConfig;
+}